Moved `node-jose` from dev to regular rependencies
`node-jose` is used for import in `src/jose/IsomorphicJoseUtility.ts`. This leads to errors when running test suites in consuming libraries, such as `@comunica/actor-http-solid-auth-fetch`. Given that this package aims at being isomorphic, maybe the direct dependency on a Node-specific package was avoided on purpose, and the issue comes from `ts-jest`, in which case this PR should be discarded.