Fixes regression on the mangled headers issue
The original issue here is that some headers are not properly passed by `solid-auth-fetcher` to the underlying library that performs the fetch. This bug was encountered when upgrading LDflex-Comunica, which it turns out passes a slightly non-standard `Headers` object. An initial fix to this issue failed to consider the execution in a Node environment, and was assuming some features only implemented by the browser. Fixing this new issue introduced a regression regarding the original problem, and this (dirty) fix aims at satisfying both use cases.