Disabled Whitesource CI status check
There is a bug in Whitesource that makes it impractical for CI status checks. However, this change still runs a Whitesource scan on push to master, so that our dashboard stays in sync with the state of the repository. It won't affect the CI pipeline, and might introduce false positives that we can manually dismiss, but it still provides some feedback on security issues.