Remove browser from package.json
The browser field hints to bundlers like webpack to use that file for
browser packages; as this is the npm package, it's unnecessary and
causes bundled code to act differently from node.js code.

Users who want to use the browser bundle can still build the project and
drop browserDist into their projects.