Interface refactor (#175)
This massively refactors the interface following conversation in #28. The features of solid-auth-fetcher are now accessed via two flavours of session API: one for single sessions, and the other for multi-sessions (more suited for a server-based approach). This also adds a new storage system: instead of storing things in the session storage, which is shared with the browser, secrets are stored in-memory.

Co-authored-by: Nicolas Seydoux <nseydoux@inrupt.com>