Restructures the documentation to have smaller pages (#233)
* Restructures the documentation to have smaller pages

All of the documentation was in a single README page, which made it harder to navigate. This splits the documentation into smaller pieces, and removes the elements not meant for beta.

* Update documentation and example code snippets

The documentation is now aligned with the new API, and is restrained to the features available for beta. The associated code snippets are up-to-date.

* Clean up API reference

The API reference contained references to unimplemented functions (e.g. `Session::init`), and mentionned unimplemented behaviours (e.g. automating redirect handling in browser).

* Update README.md

* Mention console apps on roadmap

* Adjust phrasing for colid-client-authn-node

* Update api.md

* Update browser.md

Co-authored-by: Pat McBennett <patm@inrupt.com>