Fix client registration issue in the auth code flow on NSS (#258)
For some reason, dynamic client registration information isn't persisted when interacting with NSS if stored in secure storage. If stored in insecure storage though, it works and prevents us from having to deal with the legacy token. Since the auth code flow is fixed with NSS, the workaround forcing the implicit flow is no longer necessary.