Clears both local and in-memory storage on logout (#263)
Since pieces of information are stored in both local and in-memory storage, both should be cleared on logout. This enables to log in a different IdP after logging out. This also adds a non-regression test for this bug.

To ease the non-regression test, this PR also make storage mock more flexible and without side-effect: Instead of returning a constant reference, the storage is mocked by putting the stored map in a closure returned by the mockStorageUtility function.