Fix ESLint setup
The path passed to ESLint was src/**/*.ts, not "src/**/*.ts". In
other words, expanding the wildcards was left to the shell, rather
than used as a glob in ESLint - which caused it not to actually run
on the intended files.

This resulted in a number of unfixed linting errors, that I've now
fixed.