Move DI to browser module without code duplication
TSyringe requires the class for the dependency injection to be in the same module as the dependency container. That's a problem we couldn't solve by having some classes having tsyringe decorators in several module. To resolve this without code duplication, the class that is decorated in the browser module extends the class that contains the actual code in the core module.