Clear local storage on login (#289)
If the user did not log out, and tries to login in a different tab, data stored in local storage prevents expected behaviour, and the login cannot happen. This fixes the issue by clearing the storage on login as well as on logout.

In order to not actually `logout` when logging in, but rather to simply clean up the previous session, the session cleanup logic has been moved to `SessioninfoManager`. As a side-effect, the test code has been updated to mock the SessionManagerInfo along with the StorageUtility to make tests more self-descriptive.

Co-authored-by: Vincent <Vinnl@users.noreply.github.com>