Interfaces all moved into core (#303)
* Interfaces all moved into core

* WIP - browser package test failures depending on mocks in core

* TEMPORARY: Some dependencies are resolved in the core

This does not seem to resolve the issue, but might be a first stab at it.

* Move DI to browser module without code duplication

TSyringe requires the class for the dependency injection to be in the same module as the dependency container. That's a problem we couldn't solve by having some classes having tsyringe decorators in several module. To resolve this without code duplication, the class that is decorated in the browser module extends the class that contains the actual code in the core module.

* Moved jest from devdependencies to dependencies

Since mocks are exported, and they use jest, it must be in the dependencies, not only the devdependencies.

* WIP: Start removing jest dependency from exported mocks

* Cleaned up some mock usage, moved more interfaces to Core

* Fix bad import from conflict resolution

* Fix linting issues

* Fix moved import

* Moved SessionInfoManager interfaces down to Core

* Added rollup dev deps to root package.json

* StorageUtility mocked tests fixed up now

* Updated package-lock.json with Rollup dev deps

Co-authored-by: Nicolas Seydoux <nseydoux@inrupt.com>