Refactored crypto/DPoP to be available as standalone functions
In order to build a fetch function with a token in its scope, it's easier not to depend on the DI framework to be able to generate the DPoP header, but to be able to simply call a function and to provide it the DPoP key.
This refactoring makes this possible, while remaining backward-compatible with the classes relying on DI to access crypto functions.