Fix Header management issue (#375)
If Headers is not imported from "cross-fetch", but defaults to the lib.dom Headers, then when not in a test environment, the Headers object is undefined. Adding a test for this is tricky, because the problem does not happen in the test environment. For the time being, treating Headers as plain objects is the only approach yielding consistent behaviours accross environments.