Feat/oidc dynamic client registration (#380)
Refactored code to move dcr to oidc module: Instead of having types declared in the core module, and code in the browser module, all the oidc-specific code regarding dynamic client registration moved to the oidc module. At this point, the tests have not been updated.

NB: Many options aren't tested yet, but the minimal options and expected behaviours are covered.

To have the type declarations and interfaces available in all the environment-specific modules, they were moved from oidc-client-dpop-browser back to the core module. That means that the oidc-client-dpop-browser depends on a packages that was initially meant to be solid-specific, making it less of a "pure OIDC" library, but the cost is smaller than having an independant package exporting pure OIDC types at this stage.