Moved key generation to oidc-client-dpop-browser
The DPoP key is now generated in the oidc-client-dpop-browser module, and the solid-client-authn-browser modules only calls a high-level function to get a new key.
The key is currently still stored in memory, which we want to change eventually. Also, the crypto libraries relied upon to generate the key require some node builtins, which are provided by rollup plugins for now.