Moved key generation to oidc-client-dpop-browser (#396)
The DPoP key is now generated in the oidc-client-dpop-browser module, and the solid-client-authn-browser modules only calls a high-level function to get a new key.
The key is currently still stored in memory, which we want to change eventually. Also, the crypto libraries relied upon to generate the key require some node builtins, which are provided by rollup plugins for now.

The RSA key isn't used in solid-client-authn-browser, and its generation is taking a bit longer, which fails some tests. For the time being, it has been removed from the codebase.