Support requesting a DPoP token
This integrates the DPoP support implemented in the oidc module into the browser module. For consistency, the bearer token exchange is moved to the oidc module as well, so that obtaining either type of token "feels" the same.
After this, the TokenRequester class could be removed, but it will be done in a different PR so that it appears in the git history, so that it is easier to look for.