Add a fallback redirect handler
In order to handle redirect URLs where the expected parameters cannot be found, a fallback handler is now added to the list of potential handlers. This is more aligned with the handler pattern, and prevents the auth code handler to override the implicit handler.
The URL parsing now defaults to the browser one, url-parse wasn't parsing properly the received URLs.