Workaround react components issues with localstorage
This is a stopgap solution, until a cleaner fix is found. For some reason, the functions used to clear the local storage that are working in the demo app aren't working for the react components, which is a problem for podbrowser. This enforces a clean slate, preventing previous client ids to be reused.