Gracefully handle redirect for DPoP (#493)
Since a DPoP header is scoped to one target IRI, if the request is redirected, the DPoP header is invalid, which results in a 401 or 403. To fix the issue, the fetch captures the redirection, and replays the request to the actual target IRI. An end-to-end test is associated to this behaviour.