Hoisted devDependencies using `lerna link convert`
Working on #517 showed that updating webpack on the core module broke the build, while the core module should not have dependend on webpack in the first place. This hoists all the dev dependencies to the top module, where only one version of each dependency will be used. Incidently, it resolves the issue in #517.