Feat/non oidc node (#593)
* Ported the non-OIDC code of browser module

This ports the code which does not rely on an OIDC library from soilc-client-authn-browser. `window.fetch` is replaced by "cross-fetch", and the jest environment is node (not jsdom).

Note that the core of the code is missing, and will be implemented in (hopefully) smaller PRs using a dedicated node OIDC library.