Request an authorization code with openid-client
This uses the openid-client library to request an authorization code as part of the OIDC auth code grant.
Information required to process the code after redirect, and request the token, is stored.

This commit also includes two small corrections in other OIDC handlers that prevented the tests to run now that the AuthCodeWithPkce handler does not just throw.