WIP Support for jose@3
jose v3 brings great things, such as browser API support. However, the mixing of ES modules and CJS modules wreaks a little havoc on the test environment. Supporting all of this should be considered in a future release, but should be planned for accordingly. Ideally, do not squash this commit, to come back to it later.