Token exchange after redirect
This retrieves data from storage (including an auth code), and exchanges said auth code for tokens at the issuer's token endpoint.
Both DPoP and Bearer tokens may be retrieved, as well as refresh tokens if the IdP returns one.

NB: currently, the DPoP token isn't used to build a DPoP fetch, this will be done in an upcomig PR.