Refreshes the tokens using openid-client
This implements the refresh token flow, based on openid-client. The OIDC context is retrieved from storage, and used to get a new access token from a refresh token.
This PR also trims some code that was duplicated in the AuthCodeRedirectHandler, based on the mock examples used for the refresh flow.
Note that at this point, the token refreshed isn't "used" anywhere, and still needs to be fed into the fetch function by the AuthCodeRedirectHandler.