Bearer-authenticated fetch supports the refresh token flow (#660)
This ensures that the authenticated fetch built after having received a Bearer token supports the refresh token flow. On auth failure, the fetch function tries to update its current access token, and retries the request. The refresh token is also rotated if a new token is sent by the issuer.

The DPoP fetch will incorporate a similar logic in an upcoming PR.