Refactored login function so that it can return an authenticated fetch
This changes the internal API of the fetch function: in the case of a refresh flow when the refresh token is provided (and later on in the client credential flow), getting the access token
happens directly in the back channel, and no front channel interaction is required, which means no redirection happens. So far, our `login` function only ever returned `void`, because it
was assumed that we were going through the auth code flow or the implicit flow, and that the refresh flow only happened after these once the fetch function is already instanciated.

Now, the login function can return the result of the token exchange if appropriate, and undefined if we expect the login to be completed after a redirection.