Ensured parameters are passed appropriately, and fixed test coverage
This makes sure that parameters such as the refresh token are passed all the way down the call stack. Also, this introduces something that will need to be fixed later: the OIDC login handler no longer checks for the presence of a redirect URI (which isn't necessary in the refresh token case).