Remove depdendency on existing data in e2e tests
The end-to-end tests attempt to fetch both public and private
Resources, and depended on a Pod being pre-initialised with known
access permissions.

To make the tests work with any Pod for which the credentials are
provided, we now use the Pod's WebID document as a public resource
(as that is required by the Solid spec to be public), and the Pod
root as a private resource (as that will always exist and usually
not be initialised to be public).