Add standalone versions of Session methods
These instantiate a single "default session" behind the scenes that
they all operate on. It currently also exposes a
`getDefaultSession()` function to obtain that Session, because
there is no good other way to obtain session info like the current
user's WebID. It might make sense to extend the `login` event to
supply that information.