Added environment entry in issue template (#908)
Since the repository hosts multiple packages, and the issue tracker is
at the repo level, issues would be more helpful if they specified which
package exactly is impacted by a bug.

Based on a very good point in the review comments, instead of asking
about the package impacted by the reported bug, we now ask the
environment in which the bug is observed. That will give us the same
information, but it's more user friendly.