Add temporary workaround for staying logged in
Unfortunately, regular sessions are lost when the user refreshes
the page or opens a new tab.
While we're figuring out the API for a longer-term solution, as a
temporary workaround some *resource* servers set a cookie that
keeps the user logged in after authenticated requests, and expose
the fact that they set it on a special endpoint.
After login, we store that fact in LocalStorage. This means that we
can now look for that data, and if present, indicate that the user
is already logged in.
Note that there are a lot of edge cases that won't work well with
this approach, so it will be removed in due time.

The type also still needs to be aligned with the write-side.