Write resource server session information in local storage
When logging in an Identity Provider, this tries to query an endpoint on the Resource Server hosting the user's WebID. This endpoint is expected to set a cookie for the browser, which enables authenticated requests after the user refreshes the page. 

Only one session will be active in the browser at a time, so instead of
indexing sessions with WebIDs, we only expect to ever have one active
WebID at a time.