Read the temporary session-workaround-cookie (#921)
Unfortunately, regular sessions are lost when the user refreshes
the page or opens a new tab.
While we're figuring out the API for a longer-term solution, as a
temporary workaround some *resource* servers set a cookie that
keeps the user logged in after authenticated requests, and expose
the fact that they set it on a special endpoint.
After login, we store that fact in LocalStorage. This means that we
can now look for that data, and if present, indicate that the user
is already logged in.
Note that there are a lot of edge cases that won't work well with
this approach, so it will be removed in due time.

Note that the resource server cookie isn't actually set at the session endpoint:
an authenticated request must first been sent to any "actual" resource.

If credentials: "include" is not set on the fetch call, the Set-Cookie
response is ignored by the browser, which is unfortunate, because an
authenticated request is the only way to set the cookie in the first
place. By default, fetch requests don't include browser cookies. This makes sure
that if a cookie has been set by the resource server, it is used by
`solid-client-authn-browser`.

Co-authored-by: Pat McBennett <patm@inrupt.com>
Co-authored-by: Nicolas Seydoux <nseydoux@inrupt.com>