Refactor get, implement set
The get method previously parsed the file on each read. This changes it
to only parse the file once on load, and to read it from memory on
subsequent reads.

The set method writes a new value in the in-memory cache, and persists
it on the disk. This way, the persisted value is always the latest.