De-duplicate storage API on NodeJS, improve filesystem storage
- The reason why we accept two different storages is browser-specific. It
doesn't really make sense to apply the same interface in the Node
environment, especially given that the naming of the variables
(secureStorage and insecureStorage) don't apply for NodeJS.
- This also introduces a `FileSystemStorage` class in the demo application
for authenticated Node script.  It enabled cleaning up some related bugs.
Instead of clearing the storage on login, we use it in order to silently
go through the refresh flow. If client information and the refresh token
are available on login, we can directly use them to get an access token.

Co-authored-by: Vincent <Vinnl@users.noreply.github.com>