Implemented getStoredSession
getStoredSession is a function that will return a Session based on its
session ID if the provided ID exists in storage. It is mostly useful to
handle multiple sessions on the server side, where the session object
will not always be available in the scope of the function that processes
some data.