Enforce usage of a default storage singleton
When not specifying a storage for
getClientAuthenticationWithDependencies, a default storage is
initialized. However, if we want users to be able to retrieve a session
from the default storage in node, this default storage should be a
singleton shared between all instances, and not a storage that gets
reinitialized each time a session is searched for.