Retrieve WebID from stored RS cookie
Instead of retrieving the WebID from the session information, which
aren't avalaible when the session has been restored in a new tab,
it is retrieved from local storage in the temporary RS cookie fix.

The absence of the WebID prevented the cookie workaround to be cleared,
forcing a session active in multiple tabs to remain open.