Store session ID instead of issuer
This changes the code from d4ee20b334daad9b336210bdada279f6ad5024f9.
Previously, on login, the current issuer was stored. Instead, the
current session ID is stored, which enables tracing back to the issuer,
but also to all the other session information, and in particular the ID
token, that must be validated.