The client ID is now part of the session info
The audience is a necessary part of ID token validation, and the
audience of an ID token is the client ID. That is why we now retrieve
the client ID from storage as part of the session information. Moreover,
the client ID is now exposed as part of the session information, for
convenience of the calling app.