Verify ID token when looking up current issuer
This adds the method `getCurrentIssuer` to the ClientAuthentication class.
Based on the current session ID in storage, the associated ID token is
validated, before returning its issuer to the caller. This way, we ensure
that the ID token hasn't been tampered with, that its audience is the
client associated to the current session, and that the issuer hasn't
been changed in storage either.