Verify ID token when looking up current issuer  (#1065)
* Store session ID instead of issuer: This changes the code from d4ee20b334daad9b336210bdada279f6ad5024f9.
Previously, on login, the current issuer was stored. Instead, the
current session ID is stored, which enables tracing back to the issuer,
but also to all the other session information, and in particular the ID
token, that must be validated.

* The client ID is now part of the session info: The audience is a necessary part of ID token validation, and the
audience of an ID token is the client ID. That is why we now retrieve
the client ID from storage as part of the session information. Moreover,
the client ID is now exposed as part of the session information, for
convenience of the calling app.

* Verify ID token when looking up current issuer: This adds the method `getCurrentIssuer` to the ClientAuthentication class.
Based on the current session ID in storage, the associated ID token is
validated, before returning its issuer to the caller. This way, we ensure
that the ID token hasn't been tampered with, that its audience is the
client associated to the current session, and that the issuer hasn't
been changed in storage either.

Co-authored-by: Vincent <Vinnl@users.noreply.github.com>