Store current IRI before silent authentication
This refactors code introduced in
1b4357f6b02327cf03ecf6d478392fdafc59e5ea. In the original design, the
current IRI was stored after successful login, while actually it makes
more sense to only store it during the silent authentication process,
right before silently redirecting the user away from the app, in case
client-side navigation happened and the user refreshes a page that is no
longer the one he logged in to.

For context, this is the first step to silent authentication: after a
refresh, the access token is no longer available. To log back in, the
use must be redirected to the IdP, and storing the current IRI before
redirecting makes it easier to restore it after the user is logged back
in.