Add internal support for silent authentication
This enables to specify the "prompt" option from our own internal API
all the way down to the OIDC library we are using. The "prompt" option,
when set to 'none', prevents the Identity Provider to display any screen
or to have any sort of user interaction, which is what we want for
silent authentication.

Note that the "prompt" option was not added to
the external API, because there is no reason for the calling library to
decide it needs to go through silent authentication: we make this call
entierly based on internal state. The code to implement this logic of
whether we want or not to go through silent authentication is not part
of this PR.