Get issuer canonical IRI from its own config
The user provides an issuer IRI, which allows discovering the issuer
configuration. So far, this user-provided IRI was stored on login, and
used to validate the ID token. However, this IRI may be correct, but
still differ from the issuer claim of the ID token, e.g. depending on
whether it includes a trailing slash or not. Now, the IRI that gets
stored is the canonical IRI obtained from the IdP's config, which
enforces it to be aligned with the ID token's issuer claim.