Fix: Remove PKCE-related information from DCR (#1103)
Some information only relevant to the PKCE token (which is useful to secure the client authorization and token exchange) was provided during Dynamic Client Registration, where it provided no value whatsoever. This was breaking for some Identity Providers, which are forbidding unexpected values. The DCR information is now curated.

Thanks to Wouter Termont for spotting the bug and contributing to the fix!

Signed-off-by: Wouter Termont <woutermont@gmail.com>
Co-authored-by: Wouter Termont <woutermont@gmail.com>