Removed duplication between redirectUr(i|l)
Parts of the code reads and writes the redirect URI in storage with the key `redirectUri`, and some other parts with the key `redirectUrl`. This causes confusion, and ends up with potentially creating a bug when trying to use one in silent authentication when the other is the one that has been registered as valid by the Solid Identity Provider, leading to the user being displayed an `invalid_grant` screen instead of being redirected to the app.

Technically, either `redirectUrl` or `redirectUri` are valid choices.
`redirectUrl` has been picked because it is the name of the parameter in
the login options, so picking `redirectUri` woud need either breaking
change, or to keep some room for error by having to rename the field in
the JSON object at some point. Note that the actual OAuth parameter name
is `redirect_uri`, which isn't camel cased, so a renaming would be
necessary anyways.