Store the full URI on redirect
The redirect IRI was previously stored in the session information on login. Since it is re-written upon redirect, and the redirect IRI also contains state, which may be used to restore information about the user session, the redirect IRI is only written after having been redirected from the Identity Provider. Also, the IRI was previously stored incomplete, which caused bugs on silent authentication.