Enable ESS cookie auth opt out (#1125)
* Enable ESS cookie auth opt out

So far, logging in to any Identity Provider would lead to the Resource
Server hosting the WebID being looked up for a `/session` endpoint. This
causes several issues:
- It makes additional, non-required requests if the server does not have
a session endpoint, or if the app supports silent authentication
- It has caused annoying issues when looking up the missing endpoint for
some servers
- It's a solution we want to deprecate anyways.

To resolve this, both the endpoint lookup and the cookie-based auth can
be disabled by specifying an additional option to
`handleIncomingRedirect`, `useEssSession`. It defaults to true (for
backward-compatibility), but it is considered false if silent
authentication is enabled.

Note that this uses an entry in local storage to share the state between
the Session where the option is specified and the RedirectHandler where
the session lookup actually happens. This isn't ideal, but it prevents
from changing internal APIs, only to change them back when this isn't
supported anymore.