Handle solid-oidc in client registration
This adds support for client WebID in the login process. There are a
certain number of configurations to consider: either the IdP supports
Solid-OIDC or not, and either the client provides a client_id or not.

If no client_id is provided, dynamic client registration is required,
regardless of the IdP. Most clients fall into this category today. If a
client ID is provided and it doesn't look like an IRI, or if it is an
IRI and the IdP supports Solid-OIDC, the client can be considered
already registered and there is no additional interaction required.

The library only has to be careful when the client provides a client
WebID, but the IdP soes not support Solid-OIDC.In that case, presenting
the client_id to the IdP would fail, because it would not look up the
app profile, but expect the given client_id to have previously been
registered. In that case, the client must go through DCR, and will be
considered like a public client by the resource server.

Note to reviewers: This commit only applies to `@inrupt/solid-client-authn-browser`.
The node equivalent code will be submitted in a subsequent commit.