Slower e2e tests so that NSS/Testcafé can keep up
If the tests are too fast paced, some fail to authenticate. By
trial and error, 15 seconds (!) seems to be a safe value to wait for
authentication to complete.

After investigations, it looks like the bottleneck it running firefox with
Testcafé. This is why this also sequentializes chrome and firefox runs, 
as it decreased the error rate.