Run dependabot CI without GitHub secrets
Due to
https://github.blog/changelog/2021-02-19-github-actions-workflows-triggered-by-dependabot-prs-will-run-with-read-only-permissions/,
Dependabot PRs don't have access to GH secrets. In particular, that
means they can't run e2e tests, and can't deploy a preview app on NPM.
So far, we have manually re-run the dependabot PRs to get access to the
secrets. This commit makes it so that Dependabot PRs run without e2e
tests or NPM preview, so that a human user does not need to be involved.