Prevent race condition on silent auth
Silent authentication involves a redirection. However, said redirection
happens asynchronousely, so it does not interrupt the execution flow.
This may result in the user calling `login` because
`handleIncomingRedirect` returned an unauthenticated session, while
a silent login happens in parallel. Logging in clears the storage, which
creates a race condition on storage.

This is a temporary fix, rather inelegant, which prevents
handleIncomingRedirect to return on silent authentication. This way, the
browser is redirected before handleIncomingRedirect completes, which
prevents any race condition to manifest.