Prevent race condition on login (#1243)
This fixes #1242

Silent authentication involves a redirection. However, said redirection
happens asynchronously, so it does not interrupt the execution flow.
This may result in the user calling `login` because
`handleIncomingRedirect` returned an unauthenticated session, while
a silent login happens in parallel. Logging in clears the storage, which
creates a race condition on storage.

For the the client code will block on the call until the redirect is
effective, a never resolving promise is returned.

Co-authored-by: Vincent <git@vincenttunru.com>