Add a client WebID to the demo app (#1184)
This adds a basic client WebID to the demo app. The client WebID must be
hosted somewhere where the Solid Identity Provider can look it up, which
excludes serving it under the "localhost" domain alongside the rest of
the app. We chose to have this WebID statically served by GitHub, but
alternative exist: hosting it on a Pod, or get it served statically by the
app server.

Due to the structure of the IRI, it will only be valid after the PR is
merged into the main branch.

Note to reviewers: the ESS server that supports the Solid-OIDC spec has a bug
which has been fixed but for which the fix hasn't been deployed yet,
which prevents this to work without a slight modification.