Negociate signing algorithm during DCR
Before performing DCR, the client checks the signing algorithms
supported by the OIDC issuer, and selects its preferred one. This
preferred algorithm is specified as part of the client metadata
registered to the Issuer. It is also saved to be used later during the
backchannel exchange, when verifying the ID token.