Use the signing alg when refreshing too
This makes sure that the appropriate signing algorithm is used when
refreshing the access token too, since the ID token seems to be re-sent
at this point too. Note that additionally, the grant_types are now
explicitly set at registration time, because using the offline_access
scope without having register a client with the "refresh_token" grant
does not return a refresh token.