Alg negotiation for authenticated scripts
This makes sure that if you're starting fresh with just a refresh token
and a client ID, the signing algorithm is also negociated. In contrast,
previously it was expected that the client registration stored the
signing algorithm in storage for it to be picked up when using the
refresh token.

This is not an ideal behaviour, and it is implemented this way because
of timing constraints mostly. The correct approach, that will be
implemented later, would be that the entire registered client profile
is passed in instead of just the client ID and secret, which means that
the signing algorithm would still be specified by the IdP at
registration. The current approach works, but it is more fragile.