ID token signing algorithm negotiation (#1274)
* Add support for signature negociation in core

The core package now exposes a default list of preferred signing
algorithms, as well as a function to compute which signing algorithm
supported by the OIDC Issuer is preferred by the client. The IClient
interface now also contains a field for the signing algorithm.

* Negociate signing algorithm during DCR

Before performing DCR, the client checks the signing algorithms
supported by the OIDC issuer, and selects its preferred one. This
preferred algorithm is specified as part of the client metadata
registered to the Issuer. It is also saved to be used later during the
backchannel exchange, when verifying the ID token.

* Use preferred signature in ID token verification

This uses the ID token signature algorithm that has been previously
negociated during DCR to verify the ID token.

* Throw informative error for current situation

Currently, the ID broker is missing the
'id_token_signing_alg_values_supported' property in its discovery
document. This ensures that a comprehensive error message is thrown in
this case.

* Negociate signature at DCR (browser)

This modifies the oidc and browser packages so that the ID token
signature algorithm is negociated during DCR.

* Properly parse the issuer profile

* Use the signing alg when refreshing too

This makes sure that the appropriate signing algorithm is used when
refreshing the access token too, since the ID token seems to be re-sent
at this point too. Note that additionally, the grant_types are now
explicitly set at registration time, because using the offline_access
scope without having register a client with the "refresh_token" grant
does not return a refresh token.

* Alg negotiation for authenticated scripts

This makes sure that if you're starting fresh with just a refresh token
and a client ID, the signing algorithm is also negociated. In contrast,
previously it was expected that the client registration stored the
signing algorithm in storage for it to be picked up when using the
refresh token.

This is not an ideal behaviour, and it is implemented this way because
of timing constraints mostly. The correct approach, that will be
implemented later, would be that the entire registered client profile
is passed in instead of just the client ID and secret, which means that
the signing algorithm would still be specified by the IdP at
registration. The current approach works, but it is more fragile.

Co-authored-by: Pat McBennett <patm@inrupt.com>
Co-authored-by: Vincent <Vinnl@users.noreply.github.com>