Add refresh token handling hooks
In order to allow for some use cases where using the storage isn't the
most practical solution, this makes it possible to run some custom code
when getting an access token. Two paths are considered:
- when initially logging in, when the refresh token is returned as a
response from the token endpoint. In this case,
`session.handleincomingRedirect` accepts a new parameter,
`handleRefreshToken`, which allows the caller to run custom code at this
point. An example usage for this would be to replace the current
behaviour of `generate-oidc-token`, which looks into the storage to get
the refresh token, and which could instead use this handler.
- when using a refresh token to get a new access token. In this case,
the refresh token may be rotated by the identity provider. The
`session.login` function takes for this use case a new parameter,
`handleRefreshTokenRotation`, which allows the caller to execute custom
code to manage the new refresh token, e.g. in the case the used storage
isn't adapted to handling the token. This would be useful to us in the
case of end-to-end tests, where we could share the new refresh token to
other tests via an environment variable.