Only provide ref. token handler in constructor
This changes the API from having to inject the refresh token handler in
two different places (login and handleIncomingRedirect) to only
providing the handler once, in the Session constructor. The Session
object doesn't really change behaviour, and still provides the handler
in both contexts anyways, but it's less hassle for the user.